[Cocoa] Web Inspector: Service workers can't be inspected in Safari Technology Preview https://bugs.webkit.org/show_bug.cgi?id=228001
Reviewed by Per Arne Vollan.

For inspecting targets in the WebProcess, like Service Workers, we read the `ShowDevelopMenu` default on Cocoa
platforms to determine if remote inspection should be allowed, but the sandbox broker bundle identifier used to
read that preference is different for Safari Technology Preview. We need to read the preference with the
appropriate sandbox broker bundle identifier when the host application is Safari Technology Preview.

* UIProcess/Cocoa/WebProcessProxyCocoa.mm:
(WebKit::WebProcessProxy::shouldEnableRemoteInspector):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@280039 268f45cc-cd09-0410-ab3c-d52691b4dbfc