[iOS] Tapping images on twitter.com should not toggle tab bar visibility https://bugs.webkit.org/show_bug.cgi?id=228086 rdar://79667219
Reviewed by Geoffrey Garen.

Treat taps over elements that are explicitly marked draggable and also respond to click events as "meaningful"
with respect to the private UIDelegate method `-_webView:didTapAtPoint:withResult:`. In doing this, we use the
combination of responding to drags and clicks as a proxy for content that probably results in some meaningful
interaction when tapped.

Test: fast/events/ios/meaningful-click-when-tapping-draggable-content.html

* WebProcess/WebPage/ios/WebPageIOS.mm:
(WebKit::isProbablyMeaningfulClick):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@280057 268f45cc-cd09-0410-ab3c-d52691b4dbfc