REGRESSION (r278702): Cannot login to appaloosa-store.com/users/sign_in https://bugs.webkit.org/show_bug.cgi?id=228096 <rdar://80596391>
Reviewed by Alex Christensen.

Source/WebCore:

r278702 added an optimization to FetchBodyConsumer::takeAsBlob() to avoid a copy of the data.
What I didn't realized when I wrote this optimization is that FetchBodyConsumer is copy-constructible
and its copy constructor gets called when calling FetchResponse.clone(). The copy constructor only
does a shallow-copy of its internal buffer so several FetchResponse objects can end up with their
FetchBodyConsumer using the same underlying SharedBuffer object. When that SharedBuffer is shared,
calling takeData() on it is unacceptable as it will disturb the body of other Fetch responses.

To address the issue, we now only call SharedBuffer::takeData() and avoid the copy when the
SharedBuffer has a RefCount of 1, meaning that is it not actually shared. In the cases where it
is shared, we copy the data, like we used to do before r278702.

Test: http/tests/fetch/response-clone-blob.html

* Modules/fetch/FetchBodyConsumer.cpp:
(WebCore::FetchBodyConsumer::takeAsBlob):

LayoutTests:

Add layout test coverage.

* http/tests/fetch/response-clone-blob-expected.txt: Added.
* http/tests/fetch/response-clone-blob.html: Added.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@280085 268f45cc-cd09-0410-ab3c-d52691b4dbfc