Removed incorrect test from webgl tests. https://bugs.webkit.org/show_bug.cgi?id=220720
Patch by Eleni Maria Stea <estea@igalia.com> on 2021-07-23
Reviewed by Kenneth Russell.

The removed test was based on the incorrect assumption that
deleted buffer names remain valid until the object they used to
refer to is destroyed. According to the WebGL spec:
https://www.khronos.org/registry/webgl/specs/1.0.3/ section:
5.14.5 isBuffer "Return false if the buffer's invalidated flag
is set", this is not true and names become invalid immediately.

* fast/canvas/webgl/oes-vertex-array-object.html:

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@280218 268f45cc-cd09-0410-ab3c-d52691b4dbfc