[GLIB] Make CORS allowlist test wait for promise resolution https://bugs.webkit.org/show_bug.cgi?id=228695
Reviewed by Michael Catanzaro.

r278456 introduced a CORS allow list test that tries to fetch a
resource using the fetch() function, which returns a promise. This is
working on GTK but in WPE the test fails as the promise is not yet
resolved by the time the test variable is checked.

This commit changes the test to actually report the promise failure
through catch() and wait for its resolution before checking the values.

* TestWebKitAPI/Tests/WebKitGLib/TestWebKitWebView.cpp:
(testWebViewCORSAllowlist):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@280528 268f45cc-cd09-0410-ab3c-d52691b4dbfc